Rename PackageController_ShowRouteTests → API_PackageControllerTests
split out PackageController_BuildsRoute_BuildInfoTests
